fix behavior of the state machine to start traj only if we receive one
Now the driver will check the intial status and wait for a trajectory,
only if there is a incoming one that we will try to start trajectory
mode and check if the drivers are then powered. After the end of the
execution of the trajectory, we stop motion and exit trajectory mode,
coming back to the not_checked state.